create a `coinduction.rs` module
