+++ src/test.rs
@@ -113,6 +113,7 @@ fn solve_goal(program_text: &str, goals: Vec<(&str, SolverChoice, &str)>) {
 }
 
 mod coherence_goals;
+mod coinduction;
 mod cycle;
 mod implied_bounds;
 mod impls;
@@ -388,85 +389,6 @@ fn auto_trait_with_impls() {
             forall<T> {
                 Vec<T>: Send
             }
-        } yields {
-            "No possible solution"
-        }
-    }
-}
-
-#[test]
-fn coinductive_semantics() {
-    test! {
-        program {
-            #[auto] trait Send { }
-
-            struct i32 { }
-
-            struct Ptr<T> { }
-            impl<T> Send for Ptr<T> where T: Send { }
-
-            struct List<T> {
-                data: T,
-                next: Ptr<List<T>>
-            }
-        }
-
-        goal {
-            forall<T> {
-                List<T>: Send
-            }
-        } yields {
-            "No possible solution"
-        }
-        goal {
-            forall<T> {
-                if (T: Send) {
-                    List<T>: Send
-                }
-            }
-        } yields {
-            "Unique"
-        }
-
-        goal {
-            List<i32>: Send
-        } yields {
-            "Unique"
-        }
-
-        goal {
-            exists<T> {
-                T: Send
-            }
-        } yields {
-            "Ambiguous"
-        }
-    }
-}
-
-#[test]
-fn mixed_semantics() {
-    test! {
-        program {
-            #[auto] trait Send { }
-            trait Foo { }
-
-            struct Bar { }
-
-            impl Send for Bar where Bar: Foo { }
-            impl Foo for Bar where Bar: Send { }
-        }
-
-        // We have a cycle `(T: Send) :- (T: Foo) :- (T: Send)` with a non-coinductive
-        // inner component `T: Foo` so we reject it.
-        goal {
-            Bar: Send
-        } yields {
-            "No possible solution"
-        }
-
-        goal {
-            Bar: Foo
         } yields {
             "No possible solution"
         }
